refactor(infra/require): remove the unused files && split require into different module && write a simple test
