+++ infrastructure/require/test/code.js
@@ -1,0 +1,6 @@
+module.exports = {
+  text: 'Hello custom require',
+  func() {
+    return this.text
+  }
+}
