+++ infrastructure/require/wrapper.js
@@ -1,0 +1,8 @@
+const wrapper = [
+  '(function (exports, require, module) { ',
+  '\n});'
+]
+
+module.exports = function wrap(script) {
+  return `${wrapper[0]}${script}${wrapper[1]}`
+}
