--- infrastructure/tiny-require/.eslintrc.yml
@@ -1,17 +1,0 @@
-env:
-  es6: true
-  node: true
-extends: 'eslint:recommended'
-rules:
-  indent:
-    - error
-    - 2
-  linebreak-style:
-    - error
-    - unix
-  quotes:
-    - error
-    - single
-  semi:
-    - error
-    - never
