--- infrastructure/tiny-require/README.md
@@ -1,2 +1,0 @@
-<h1>__local_require__</h1>
-[CommonJS] Wrap code &amp;&amp; bind code as a module into global module object.
