--- infrastructure/tiny-require/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "local-require",
-  "version": "1.0.0",
-  "description": "[CommonJS] Wrap code &amp;&amp; bind code as a module into global module object.",
-  "main": "index.js",
-  "scripts": {
-    "test": "jest"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/loatheb/__local_require__.git"
-  },
-  "keywords": [
-    "require"
-  ],
-  "author": "loatheb",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/loatheb/__local_require__/issues"
-  },
-  "homepage": "https://github.com/loatheb/__local_require__#readme",
-  "devDependencies": {
-    "eslint": "^3.19.0"
-  },
-  "dependencies": {
-    "is-native-modules": "^1.0.3"
-  }
-}
